+++ RELEASENOTES.md
@@ -11,6 +11,9 @@ This release of SpriteSomething includes many basic features of the main design 
 ### Super Metroid
 
 * Add Spiffy Button for toggling Screw Attack animation split (new Spin Attack without Space Jump graphics) that is honored on-injection
+* Check for VT Tournament flags
+* Check for Combo Tournament flags
+* Check for VARIA Tournament flags
 
 ## Features